+++ src/sdk/Cryptography/Hint/Components/ComponentHintScanner.ts
@@ -3,8 +3,10 @@ import {
 } from "../../Helper/ResultCollector/Component/ComponentCryptographyResultColletor";
 import { BaseCryptographyScanner } from "../../BaseCryptographyScanner";
 import { HintsInRangeResponse } from "../../../Clients/Cryptography/ICryptographyClient";
-import { CryptographyClient } from "../../../Clients/Cryptography/CryptographyClient";
 import { Component } from "../../../types/common/types";
+import {
+  CryptographyHttpClient
+} from "../../../Clients/Cryptography/CryptographyHttpClient";
 
 /**
  * Scanner for detecting cryptographic hints in software components.
@@ -25,7 +27,7 @@ export class ComponentHintScanner
    * @returns {HintsResponse} A promise that resolves to a HintsResponse containing detected cryptographic hints.
    */
   public async scan(req: Component[]):Promise<HintsInRangeResponse>{
-    const cryptographyClient = new CryptographyClient(
+    const cryptographyClient = new CryptographyHttpClient(
       this.config.API_KEY, // API KEY
       this.config.API_URL, // Destination Host
       this.config.GRPC_PROXY, // Proxy Host